Protocol update.

+++ movement.h
@@ -17,8 +17,8 @@ void movement_stop(uint8_t reason);
 // start executing queued movements
 void movement_start();
 
-// push movement to queue, returns -1 on queue full
-int8_t movement_push(struct movement_block * op);
+// push movement to queue, returns -1 on queue full, otherwise returns tag
+int16_t movement_push(struct movement_block * op);
 
 // set movement speeds immediately, returns -1 if queue started
 int8_t movement_jog(struct movement_block * op);
